remove limitAndoffset
