+++ fe/fe-core/src/main/antlr4/org/apache/doris/nereids/DorisParser.g4
@@ -122,9 +122,9 @@ sortItem
     ;
 
 limitClause
-    : (LIMIT INTEGER_VALUE)?
-    | (LIMIT INTEGER_VALUE OFFSET INTEGER_VALUE)?
-    | (LIMIT INTEGER_VALUE COMMA INTEGER_VALUE)?
+    : (LIMIT limit=INTEGER_VALUE)?
+    | (LIMIT limit=INTEGER_VALUE OFFSET offset=INTEGER_VALUE)?
+    | (LIMIT offset=INTEGER_VALUE COMMA limit=INTEGER_VALUE)?
     ;
 
 joinType
