--- fe/fe-core/src/main/java/org/apache/doris/nereids/properties/LimitAndOffset.java
@@ -1,56 +1,0 @@
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-package org.apache.doris.nereids.properties;
-
-/**
- * limit clause, represents the limit and offset
- * select * from t limit 5; //limit only
- * select * from t order by a limit 5, 100 // limit=5 and offset=100
- * select * from t order by a limit 5 offset 100 // limit=5 and offset = 100
- * `offset` must be used with `order by`
- */
-public class LimitAndOffset {
-    private long limit = 0;
-
-    private long offset = 0;
-
-    public LimitAndOffset(long limit, long offset) {
-        this.limit = limit;
-        this.offset = offset;
-    }
-
-    public long getOffset() {
-        return offset;
-    }
-
-    public void setOffset(long offset) {
-        this.offset = offset;
-    }
-
-    public long getLimit() {
-        return limit;
-    }
-
-    public void setLimit(long limit) {
-        this.limit = limit;
-    }
-
-
-
-
-}
