+++ fe/fe-core/src/main/java/org/apache/doris/nereids/trees/plans/logical/LogicalLimit.java
@@ -72,7 +72,7 @@ public class LogicalLimit<CHILD_TYPE extends Plan> extends LogicalUnary<CHILD_TY
 
     @Override
     public String toString() {
-        return "LogicalLimit (" + offset + "," + limit + ")";
+        return "LogicalLimit ( offset=" + offset + ", limit=" + limit + ")";
     }
 
     @Override
